save token in client side

+++ server/controllers/createUser.js
@@ -5,6 +5,10 @@ export default async function createUser(path,newUser){
       try {
         const data = await readFile(path);
         users = JSON.parse(data);
+        const userExist = users.filter(user => user.userName === newUser.userName)
+        if(userExist.length){
+            throw new Error("user name exist")
+        }
         users.push(newUser)
         await writeFile(path,JSON.stringify(users))
       } catch (error) {
